REFACTOR: moved class `BaseController` to `./common` namespace: - excluded it from barrelsby config, becuase it should not be used alone.
