+++ src/controllers/api/AccountsController.ts
@@ -19,7 +19,7 @@ import {
   Status,
   Summary
 } from "@tsed/schema";
-import { BaseController } from "./BaseController";
+import { BaseController } from "./common/BaseController";
 
 const d = new OASDocs("account");
