+++ src/controllers/api/PenalitiesController.ts
@@ -21,7 +21,7 @@ import {
   Status,
   Summary
 } from "@tsed/schema";
-import { BaseController } from "./BaseController";
+import { BaseController } from "./common/BaseController";
 
 const FIX_AMNT = "fix-amount";
 const RATE_OBS = "rate-of-base";
